+++ src/frequenz/sdk/config/_logging_actor.py
@@ -4,16 +4,14 @@
 """Read and update logging severity from config."""
 
 import logging
-from collections.abc import Mapping
 from dataclasses import dataclass, field
-from typing import Annotated, Any
+from typing import Annotated, Sequence, assert_never
 
 import marshmallow
 import marshmallow.validate
-from frequenz.channels import Receiver
 
 from ..actor import Actor
-from ._util import load_config
+from ._manager import ConfigManager, wait_for_first
 
 _logger = logging.getLogger(__name__)
 
@@ -84,26 +82,13 @@ class LoggingConfigUpdatingActor(Actor):
 
         ```python
         import asyncio
-        from collections.abc import Mapping
-        from typing import Any
 
-        from frequenz.channels import Broadcast
-        from frequenz.sdk.config import LoggingConfigUpdatingActor, ConfigManagingActor
+        from frequenz.sdk.config import LoggingConfigUpdatingActor
         from frequenz.sdk.actor import run as run_actors
 
         async def run() -> None:
-            config_channel = Broadcast[Mapping[str, Any]](name="config", resend_latest=True)
-            actors = [
-                ConfigManagingActor(
-                    config_paths=["config.toml"], output=config_channel.new_sender()
-                ),
-                LoggingConfigUpdatingActor(
-                    config_recv=config_channel.new_receiver(limit=1)).map(
-                        lambda app_config: app_config.get("logging", {}
-                    )
-                ),
-            ]
-            await run_actors(*actors)
+            config_manager: ConfigManager = ...
+            await run_actors(LoggingConfigUpdatingActor(config_manager))
 
         asyncio.run(run())
         ```
@@ -112,10 +97,13 @@ class LoggingConfigUpdatingActor(Actor):
         will be updated as well.
     """
 
+    # pylint: disable-next=too-many-arguments
     def __init__(
         self,
+        config_manager: ConfigManager,
+        /,
         *,
-        config_recv: Receiver[Mapping[str, Any]],
+        config_key: str | Sequence[str] = "logging",
         log_datefmt: str = "%Y-%m-%dT%H:%M:%S%z",
         log_format: str = "%(asctime)s %(levelname)-8s %(name)s:%(lineno)s: %(message)s",
         name: str | None = None,
@@ -123,7 +111,9 @@ class LoggingConfigUpdatingActor(Actor):
         """Initialize this instance.
 
         Args:
-            config_recv: The receiver to listen for configuration changes.
+            config_manager: The configuration manager to use.
+            config_key: The key to use to retrieve the configuration from the
+                configuration manager.  If `None`, the whole configuration will be used.
             log_datefmt: Use the specified date/time format in logs.
             log_format: Use the specified format string in logs.
             name: The name of this actor. If `None`, `str(id(self))` will be used. This
@@ -135,7 +125,9 @@ class LoggingConfigUpdatingActor(Actor):
             in the application (through a previous `basicConfig()` call), then the format
             settings specified here will be ignored.
         """
-        self._config_recv = config_recv
+        self._config_receiver = config_manager.new_receiver(
+            config_key, LoggingConfig, base_schema=None
+        )
 
         # Setup default configuration.
         # This ensures logging is configured even if actor fails to start or
@@ -149,21 +141,45 @@ class LoggingConfigUpdatingActor(Actor):
             datefmt=log_datefmt,
             level=logging.INFO,
         )
-        self._update_logging(self._current_config)
+        _logger.info("Applying initial default logging configuration...")
+        self._reconfigure(self._current_config)
 
     async def _run(self) -> None:
         """Listen for configuration changes and update logging."""
-        async for message in self._config_recv:
-            try:
-                new_config = load_config(LoggingConfig, message)
-            except marshmallow.ValidationError:
-                _logger.exception(
-                    "Invalid logging configuration received. Skipping config update"
-                )
-                continue
+        self._reconfigure(
+            await wait_for_first(
+                self._config_receiver, receiver_name=str(self), allow_none=True
+            )
+        )
+        async for config_update in self._config_receiver:
+            self._reconfigure(config_update)
 
-            if new_config != self._current_config:
-                self._update_logging(new_config)
+    def _reconfigure(self, config_update: LoggingConfig | Exception | None) -> None:
+        """Update the logging configuration.
+
+        Args:
+            config_update: The new configuration, or an exception if there was an error
+                parsing the configuration, or `None` if the configuration was unset.
+        """
+        match config_update:
+            case LoggingConfig():
+                _logger.info(
+                    "New configuration received, updating logging configuration."
+                )
+                self._update_logging(config_update)
+            case None:
+                _logger.info(
+                    "Configuration was unset, resetting to the default "
+                    "logging configuration."
+                )
+                self._update_logging(LoggingConfig())
+            case Exception():
+                _logger.info(
+                    "New configuration has errors, keeping the old logging "
+                    "configuration."
+                )
+            case unexpected:
+                assert_never(unexpected)
 
     def _update_logging(self, config: LoggingConfig) -> None:
         """Configure the logging level."""